backend/requirements.txt
fastapi
pyyaml
openai
azure-identity
uvicorn
langchain-openai
langchain
langchain-community
requests
beautifulsoup4
yfinance
pandas
numpy
python-dateutil
aiohttp
asyncio
